+++ src/main/kotlin/eu/timerertim/knevo/neat/InnovationCounter.kt
@@ -1,4 +1,4 @@
-package eu.timerertim.kneat
+package eu.timerertim.knevo.neat
 
 object InnovationCounter {
