+++ src/main/kotlin/eu/timerertim/knevo/neat/NodeGene.kt
@@ -1,4 +1,4 @@
-package eu.timerertim.kneat
+package eu.timerertim.knevo.neat
 
 /**
  * NodeGene represents the nodes of the neural network
