+++ src/main/kotlin/eu/timerertim/knevo/neat/config/Seed.kt
@@ -1,4 +1,4 @@
-package eu.timerertim.kneat.config
+package eu.timerertim.knevo.neat.config
 
 import kotlin.random.Random
